requirements.txt
django==4.0.6
psycopg2-binary==2.9.3
gunicorn==20.1.0
Pillow==9.2.0
django-modeltranslation==0.18.2
celery==5.2.7
beautifulsoup4==4.11.1
redis==4.3.4
django-environ==0.9.0